bundle_requirements.txt
argparse==1.2.1
zc.buildout==2.1.0
clint==0.3.1
pyjip==0.4
idxtools==0.9.1
